Changed exceptions from SNMP tree to consolidate down to one OID not found exception. Index based functions no longer check bounds or throw checked exceptions. General code cleanup and refactoring.

+++ src/main/java/com/sppad/jots/util/SnmpUtils.java
@@ -46,7 +46,7 @@ public class SnmpUtils
 		oidInts[0] = length;
 
 		for (int i = 0; i < length; i++)
-			oidInts[i + 1] = (int) string.charAt(i);
+			oidInts[i + 1] = string.charAt(i);
 
 		return oidInts;
 	}
